% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_utils.R
\name{my_mean}
\alias{my_mean}
\title{Compute the mean of a numeric vector}
\usage{
my_mean(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric scalar, the mean of \code{x}.
}
\description{
This function computes the arithmetic mean.
}
\examples{
my_mean(c(1, 2, 3))
}
